Tighten up tests more
