+++ forge/src/main/kotlin/com/github/kittinunf/forge/core/Serializable.kt
@@ -1,0 +1,7 @@
+package com.github.kittinunf.forge.core
+
+interface Serializable {
+
+    fun serialize(): JSON
+
+}
