+++ forge/src/test/kotlin/com/github/kittinunf/forge/BaseTest.kt
@@ -5,25 +5,24 @@ import java.io.File
 
 open class BaseTest {
 
-    val assetsDir: File
+    private val assetsDir: File
 
-    lateinit var userJson: String
-    lateinit var userFriendsJson: String
-    lateinit var usersJson: String
+    val userJson: String
+    val userFriendsJson: String
+    val usersJson: String
 
     init {
         val dir = System.getProperty("user.dir")
         val file = File(dir)
 
         assetsDir = file.walkTopDown().filter { it.name == "assets" }.first()
-    }
-
-    @Before
-    fun setUp() {
         userJson = File(assetsDir, "user.json").readText()
         userFriendsJson = File(assetsDir, "user_and_friends.json").readText()
         usersJson = File(assetsDir, "users.json").readText()
     }
+
+    @Before
+    fun setUp() {}
 
 }
