+++ forge/src/test/kotlin/com/github/kittinunf/forge/JSONMappingArrayTest.kt
@@ -6,20 +6,20 @@ import com.github.kittinunf.forge.core.JSON
 import com.github.kittinunf.forge.core.apply
 import com.github.kittinunf.forge.core.at
 import com.github.kittinunf.forge.core.map
+import com.github.kittinunf.forge.core.maybeAt
 import com.github.kittinunf.forge.core.maybeList
+import com.github.kittinunf.forge.model.Company
+import com.github.kittinunf.forge.model.Dog
+import com.github.kittinunf.forge.model.User
+import com.github.kittinunf.forge.model.UserWithCompany
+import com.github.kittinunf.forge.model.UserWithDogs
 import com.github.kittinunf.forge.util.create
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.CoreMatchers.nullValue
-import org.hamcrest.MatcherAssert.assertThat
+import org.junit.Assert.assertThat
 import org.junit.Test
 
 class JSONMappingArrayTest : BaseTest() {
-
-    data class User(val id: Int,
-                    val username: String,
-                    val name: String,
-                    val age: Int,
-                    val email: String)
 
     class UserDeserializer : Deserializable<User> {
         override fun deserialize(json: JSON): DeserializedResult<User> {
@@ -32,20 +32,18 @@ class JSONMappingArrayTest : BaseTest() {
         }
     }
 
-    data class Company(val name: String, val catchPhrase: String)
-
     val companyDeserializer = { json: JSON ->
         ::Company.create.
                 map(json at "name").
                 apply(json at "catch_phrase")
     }
 
-    data class UserWithCompany(val id: Int, val username: String, val company: Company)
 
     val userModelWithCompany = { json: JSON ->
         ::UserWithCompany.create.
                 map(json at "id").
                 apply(json at "username").
+                apply(json maybeAt "is_deleted").
                 apply(json.at("company", companyDeserializer))
     }
 
@@ -72,16 +70,12 @@ class JSONMappingArrayTest : BaseTest() {
         assertThat(companies[6].catchPhrase, equalTo("Configurable multimedia task-force"))
     }
 
-    data class Dog(val name: String, val breed: String, val male: Boolean)
-
     val dogDeserializer = { j: JSON ->
         ::Dog.create.
                 map(j at "name").
                 apply(j at "breed").
                 apply(j at "is_male")
     }
-
-    data class UserWithDogs(val email: String, val phone: String, val dogs: List<Dog>?)
 
     val userWithDogDeserializer = { j: JSON ->
         ::UserWithDogs.create.
