+++ forge/src/test/kotlin/com/github/kittinunf/forge/JSONMappingObjectTest.kt
@@ -1,6 +1,5 @@
 package com.github.kittinunf.forge
 
-import com.github.kittinunf.forge.JSONMappingObjectTest.Friend.Address
 import com.github.kittinunf.forge.core.Deserializable
 import com.github.kittinunf.forge.core.DeserializedResult
 import com.github.kittinunf.forge.core.JSON
@@ -11,19 +10,25 @@ import com.github.kittinunf.forge.core.map
 import com.github.kittinunf.forge.core.maybeAt
 import com.github.kittinunf.forge.helper.deserializeDate
 import com.github.kittinunf.forge.helper.toDate
+import com.github.kittinunf.forge.model.Company
+import com.github.kittinunf.forge.model.Friend
+import com.github.kittinunf.forge.model.Friend.Address
+import com.github.kittinunf.forge.model.SimpleUser
+import com.github.kittinunf.forge.model.User
+import com.github.kittinunf.forge.model.UserCreatedAt
+import com.github.kittinunf.forge.model.UserWithCompany
+import com.github.kittinunf.forge.model.UserWithFriends
+import com.github.kittinunf.forge.model.UserWithOptionalFields
 import com.github.kittinunf.forge.util.create
 import com.github.kittinunf.forge.util.curry
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.CoreMatchers.notNullValue
 import org.hamcrest.CoreMatchers.nullValue
 import org.hamcrest.MatcherAssert.assertThat
-import org.json.JSONObject
 import org.junit.Test
 import java.util.*
 
 class JSONMappingObjectTest : BaseTest() {
-
-    data class SimpleUser(val id: Int, val name: String)
 
     class SimpleUserDeserializer : Deserializable<SimpleUser> {
         override fun deserialize(json: JSON): DeserializedResult<SimpleUser> =
@@ -42,69 +47,14 @@ class JSONMappingObjectTest : BaseTest() {
         assertThat(ex, nullValue())
     }
 
-    data class User(val id: Int, val name: String, val age: Int, val email: String)
-
     class UserDeserializer : Deserializable<User> {
         override fun deserialize(json: JSON): DeserializedResult<User> =
                 ::User.create.
                         map(json at "id").
+                        apply(json at "username").
                         apply(json at "name").
                         apply(json at "age").
                         apply(json at "email")
-    }
-
-    @Test
-    fun testUserModelCurry() {
-        val json = JSON.parse((JSONObject(userJson)))
-
-        val curry = ::User.curry()
-
-        val id: DeserializedResult<Int> = (json at "id")
-        val name: DeserializedResult<String> = (json at "name")
-        val age: DeserializedResult<Int> = (json at "age")
-        val email: DeserializedResult<String> = (json at "email")
-
-        val user = curry(id.get())(name.get())(age.get())(email.get())
-
-        assertThat(user.id, equalTo(1))
-        assertThat(user.name, equalTo("Clementina DuBuque"))
-        assertThat(user.age, equalTo(46))
-        assertThat(user.email, equalTo("Rey.Padberg@karina.biz"))
-    }
-
-    data class UserWithOptionalFields(val name: String, val city: String?, val gender: String?, val phone: String, val weight: Float)
-
-    class UserWithOptionalFieldsDeserializer : Deserializable<UserWithOptionalFields> {
-
-        override fun deserialize(json: JSON): DeserializedResult<UserWithOptionalFields> =
-                ::UserWithOptionalFields.create.
-                        map(json at "name").
-                        apply(json maybeAt "city").
-                        apply(json maybeAt "gender").
-                        apply(json at "phone").
-                        apply(json at "weight")
-
-    }
-
-    @Test
-    fun testUserModelOptionalCurry() {
-        val json = JSON.parse((JSONObject(userJson)))
-
-        val curry = ::UserWithOptionalFields.curry()
-
-        val name: DeserializedResult<String> = (json at "name")
-        val phone: DeserializedResult<String> = (json at "phone")
-        val weight: DeserializedResult<Float> = (json at "weight")
-        val city: DeserializedResult<String> = (json maybeAt "city")
-        val gender: DeserializedResult<String> = (json maybeAt "gender")
-
-        val user = curry(name.get())(city.get())(gender.get())(phone.get())(weight.get())
-
-        assertThat(user.name, equalTo("Clementina DuBuque"))
-        assertThat(user.phone, equalTo("024-648-3804"))
-        assertThat(user.weight, equalTo(72.5f))
-        assertThat(user.city, nullValue())
-        assertThat(user.gender, nullValue())
     }
 
     @Test
@@ -129,19 +79,17 @@ class JSONMappingObjectTest : BaseTest() {
         assertThat(user.email, equalTo("Rey.Padberg@karina.biz"))
     }
 
-    data class Company(val name: String, val catchPhrase: String)
-
     val companyDeserializer = { json: JSON ->
         ::Company.create.
                 map(json at "name").
                 apply(json at "catch_phrase")
     }
 
-    data class UserWithCompany(val username: String, val isDeleted: Boolean, val company: Company)
 
     val userModelWithCompanyDeserializer = { json: JSON ->
         ::UserWithCompany.create.
-                map(json at "username").
+                map(json at "id").
+                apply(json at "username").
                 apply(json at "is_deleted").
                 apply(json.at("company", companyDeserializer))
     }
@@ -151,13 +99,13 @@ class JSONMappingObjectTest : BaseTest() {
         val result = Forge.modelFromJson(userJson, userModelWithCompanyDeserializer)
         val user: UserWithCompany = result.get()
 
+        assertThat(user.id, equalTo(1))
         assertThat(user.username, equalTo("Moriah.Stanton"))
         assertThat(user.isDeleted, equalTo(true))
         assertThat(user.company.name, equalTo("Hoeger LLC"))
         assertThat(user.company.catchPhrase, equalTo("Centralized empowering task-force"))
     }
 
-    data class UserCreatedAt(val id: Int, val createdAt: Date)
 
     class UserCreatedAt1Deserializer : Deserializable<UserCreatedAt> {
 
@@ -176,7 +124,7 @@ class JSONMappingObjectTest : BaseTest() {
     }
 
     @Test
-    fun testUserModelCreatedAtDeserializing1() {
+    fun testUserModelCreatedAt1Deserializing() {
         val result = Forge.modelFromJson(userJson, UserCreatedAt1Deserializer())
         val (user, ex) = result
 
@@ -192,7 +140,7 @@ class JSONMappingObjectTest : BaseTest() {
     }
 
     @Test
-    fun testUserModelCreatedAtDeserializing2() {
+    fun testUserModelCreatedAt2Deserializing() {
         val result = Forge.modelFromJson(userJson, UserCreatedAt2Deserializer())
         val (user, ex) = result
 
@@ -207,6 +155,18 @@ class JSONMappingObjectTest : BaseTest() {
         assertThat(c.get(Calendar.YEAR), equalTo(2015))
     }
 
+    class UserWithOptionalFieldsDeserializer : Deserializable<UserWithOptionalFields> {
+
+        override fun deserialize(json: JSON): DeserializedResult<UserWithOptionalFields> =
+                ::UserWithOptionalFields.create.
+                        map(json at "name").
+                        apply(json maybeAt "city").
+                        apply(json maybeAt "gender").
+                        apply(json at "phone").
+                        apply(json at "weight")
+
+    }
+
     @Test
     fun testUserModelWithOptionalFieldsDeserializing() {
         val result = Forge.modelFromJson(userJson, UserWithOptionalFieldsDeserializer())
@@ -217,12 +177,6 @@ class JSONMappingObjectTest : BaseTest() {
         assertThat(user.gender, nullValue())
         assertThat(user.phone, equalTo("024-648-3804"))
         assertThat(user.weight, equalTo(72.5f))
-    }
-
-    data class Friend(val id: Int, val name: String, val address: Friend.Address) {
-
-        data class Address(val street: String, val suite: String, val city: String)
-
     }
 
     class AddressDeserializer : Deserializable<Friend.Address> {
@@ -243,7 +197,6 @@ class JSONMappingObjectTest : BaseTest() {
                         apply(json.at("address", AddressDeserializer()::deserialize))
     }
 
-    data class UserWithFriends(val id: Int, val name: String, val age: Int, val email: String, val friends: List<Friend>)
 
     class UserWithFriendsDeserializer : Deserializable<UserWithFriends> {
         override fun deserialize(json: JSON): DeserializedResult<UserWithFriends> =
